Fix variables in block that were overwriting each other
Even when a task is skipped, the variable is set so it's not possible to use the same variable for multiple tasks counting on only the successful task to have the data that I want.

Update tests to run verification in a separate play so handlers are triggered before valiadation tasks are run.

Rename test sceanarios to match behavior.